IS-707 : Salesforce Delta Changes Deployment
With this latest code it will compare the branch from the pull request versus the branch that is receiving the pull request.
If there is any differences between both it will then generate the package.
The script check if there is both a package and a destructiveChanges files before attemping the deployment.
If there is none, it will add to the debug from the job but wont fail the job.